Add charges to Composition based on oxi_guesses (#974)
Added a function that returns a copy of a Composition decorated
using oxidation states determined from `oxi_state_guesses`.
Also refactored the `oxi_state_guesses` to move the computation
of likely oxidation states to a private utility function, so
that it could return both the average oxidation states for each
element and the oxidation states for each site.

As part of this modification, the `oxi_state_guesses` function
now returns a tuple of operations rather than a list. I could easily
modify the code to still return a list, but decided it was unnecessary
since both collections are iterable and addressable.