Added 'get_ir_reciprocal_mesh_map' function.
This function does the same as 'get_ir_reciprocal_mesh' but instead of returning the irreducible mesh it returns the grid and the mapping that can be used to map each reducible kpoint of the full grid to the irreducible ones. This data is often needed in physical simulations, because the you need to know which kpoints are related by symmetry.